[common] Separate string operations from Global.h
